requirements.txt
Flask
Werkzeug
python-telegram-bot
python-dotenv
gunicorn
